feat: Ignore changes to an ignored type
If a type is in the Types of an ignore file, but is not actually removed, ignore all removed members on that type instead.

In general ignoring individual members explicitly is preferable, but this option is useful in certain special cases, eg breaking changes involving large numbers of properties coming from generated code.